\name{Accident}
\docType{data}
\alias{Accident}
\title{Ship Accidents}
\description{
  a cross-section\cr

 \emph{number of observations} :  40 \cr
   \emph{observation} :   \cr
   \emph{country} :   \cr
    }
\usage{data(Accident)}
\format{A dataframe containing :
  \describe{
    \item{Type}{ship type}
    \item{TA}{type indicators }
    \item{TB}{type indicators }
    \item{TC}{type indicators }
    \item{TD}{type indicators }
    \item{TE}{type indicators}
    \item{T6064}{year constructed indicators }
    \item{T6569}{year constructed indicators }
    \item{T7074}{year constructed indicators }
    \item{T7579}{year constructed indicators}
    \item{O6074}{year operated indicators }
    \item{O7579}{year operated indicators }
    \item{Months}{measure of service amount}
    \item{Acc}{accidents}
  }
}
\source{
    McCullagh, P.  and  J.  Nelder (1983) \emph{Generalized linear methods}, New York:Chapman and Hall.. \cr
}
\references{
    Greene, W.H. (2003) \emph{Econometric Analysis}, Prentice Hall, Table F21.3. \cr
}
\section{Original description}{``40 observations on 5 types in 4 vintages and 2 service periods.''}
\keyword{datasets}