\name{Airline}
\docType{data}
\alias{Airline}
\title{Cost for U.S. Airlines}
\description{
  a panel of 6 observations  from 1970  to 1984

 \emph{number of observations} :  90

 \emph{observation} :  firms

   \emph{country} :  United States

}
\usage{data(Airline)}
\format{A dataframe containing :
  \describe{
    \item{I}{airline}
    \item{T}{year}
    \item{C}{total cost, in \$1,000}
    \item{Q}{output, in revenue passenger miles, index number}
    \item{PF}{fuel price}
    \item{LF}{load factor, the average capacity utilization of the fleet}
  }
}
\note{\dQuote{These data are a subset of a larger data set provided to the   author by Professor Moshe Kim. They were originally constructed by   Christensen Associates of Madison, Wisconsin.}}
\references{
    Greene, W.H. (2003) \emph{Econometric Analysis}, Prentice Hall, Table F7.1.
}
\section{Original description}{\dQuote{90 observations on 6 firms for 1970-1984.}}
\keyword{datasets}