\name{Conso}
\docType{data}
\alias{Conso}
\title{Consumption and Income}
\description{
  annual observations from 1970 to 1979\cr

 \emph{number of observations} :  10 \cr
   \emph{observation} :  Country \cr
   \emph{country} :  United States \cr
    }
\usage{data(Conso)}
\format{A dataframe containing :
  \describe{
    \item{C}{consumption}
    \item{Y}{disposable income}
  }
}
\source{
    Concil of Economic Advisors, Economic Report of the President   Washington D.C.: U.S. Government Printing Office, 1987. \cr
}
\references{
    Greene, W.H. (2003) \emph{Econometric Analysis}, Prentice Hall, Table F1.1. \cr
}
\section{Original description}{``10 yearly observations, 1970-1979.''}
\keyword{datasets}