\name{EconEduc}
\docType{data}
\alias{EconEduc}
\title{Program Effectiveness}
\description{
  a cross-section\cr

 \emph{number of observations} :  32 \cr
   \emph{observation} :  individuals \cr
   \emph{country} :  United States \cr
    }
\usage{data(EconEduc)}
\format{A dataframe containing :
  \describe{
    \item{OBS}{observation}
    \item{GPA}{grade point average}
    \item{TUCE}{test score on economic test}
    \item{PSI}{participation in program}
    \item{GRADE}{grade increase (1) or decrease (0) indicator}
  }
}
\source{
    Spector, L.  and  M.  Mazzeo (1980) ``Probit analysis and economic education'', \emph{Journal of Economic Education }, \bold{11}, 37-44. \cr
}
\references{
    Greene, W.H. (2003) \emph{Econometric Analysis}, Prentice Hall, Table F21.1. \cr
}
\section{Original description}{``32 cross section observations.''}
\keyword{datasets}