\name{Electric}
\docType{data}
\alias{Electric}
\title{Cost Function}
\description{
  a cross-section from 1970

 \emph{number of observations} :  158

 \emph{observation} :  firms

   \emph{country} :  United States

}
\usage{data(Electric)}
\format{A dataframe containing :
  \describe{
    \item{id}{observation}
    \item{year}{1970 for all observations}
    \item{cost}{total cost}
    \item{q}{total output}
    \item{pl}{wage rate}
    \item{sl}{cost share for labor}
    \item{pk}{capital price index}
    \item{sk}{cost share for capital}
    \item{pf}{fuel price}
    \item{sf}{cost share for fuel}
  }
}
\note{\dQuote{Note the file contains some extra observations. These are holding companies. Use only the first 123 observations to replicate Christensen and Greene.}}
\source{
    Christensen, L.  and  W. H.  Greene (1976) \dQuote{Economies of scale in U.S. electric power generation}, \emph{Journal of Political Economy }, \bold{84}, 655-676.
}
\references{
    Greene, W.H. (2003) \emph{Econometric Analysis}, Prentice Hall, chapter 4, 317-320.
}
\section{Original description}{\dQuote{123 1970 cross-section firm level observations.}}
\keyword{datasets}