\name{ExchRate}
\docType{data}
\alias{ExchRate}
\title{Boolerslev and Ghysels Exchange Rate Data}
\description{
  daily observations from  to \cr

 \emph{number of observations} :  1974 \cr
   \emph{observation} :  country \cr
   \emph{country} :  Germany \cr
    }
\usage{data(ExchRate)}
\format{A dataframe containing :
  \describe{
    \item{Y}{nominal return on Mark/Pound exchange rate, daily}
  }
}
\source{
    Bollerslev, T. (1986) ``Generalized autoregressive conditional heteroscedasticity'', \emph{Journal of Econometrics }, \bold{31}, 307-327. \cr
}
\references{
    Greene, W.H. (2003) \emph{Econometric Analysis}, Prentice Hall, Table F11.1. \cr
}
\section{Original description}{``1974 daily observations.''}
\keyword{datasets}