\name{Fair}
\docType{data}
\alias{Fair}
\title{Fair's (1977) Extramarital Affairs Data}
\description{
  a cross-section\cr

 \emph{number of observations} :  601 \cr
   \emph{observation} :  individuals \cr
   \emph{country} :  United States \cr
    }
\usage{data(Fair)}
\format{A dataframe containing :
  \describe{
    \item{ID}{observation}
    \item{X1}{not used}
    \item{X2}{not used}
    \item{Z1}{sex}
    \item{Z2}{age}
    \item{Z3}{number of years married}
    \item{Z4}{children}
    \item{Z5}{religiousness}
    \item{Z6}{education}
    \item{X3}{not used}
    \item{Z7}{occupation}
    \item{Z8}{self rating of marriage}
    \item{Y}{number of affairs in past year}
    \item{X4}{not used}
    \item{X5}{not used}
  }
}
\source{
    Fair, R. (1977) ``A note on computation of the tobit estimator'', \emph{Econometrica }, \bold{45}, 1723-1727. \cr
    \url{http://fairmodel.econ.yale.edu/rayfait/pdf/1978ADAT.ZIP}. \cr
}
\references{
    Greene, W.H. (2003) \emph{Econometric Analysis}, Prentice Hall,  Table F22.2. \cr
}
\section{Original description}{``601 cross-section observations.''}
\keyword{datasets}