\name{Gasoline}
\docType{data}
\alias{Gasoline}
\title{The U.S. Gasoline Market}
\description{
  annual observations from 1960 to 1995\cr

 \emph{number of observations} :  36 \cr
   \emph{observation} :  country \cr
   \emph{country} :  United States \cr
    }
\usage{data(Gasoline)}
\format{A dataframe containing :
  \describe{
    \item{Year}{the year}
    \item{G}{total U.S. gasoline consumption, computed as total expenditure divided by price index}
    \item{Pg}{price index for gasoline}
    \item{Y}{per capita disposable income}
    \item{Pnc}{price index for new cars}
    \item{Puc}{price index for used cars}
    \item{Ppt}{price index for public transportation}
    \item{Pd}{price index for consumer durables}
    \item{Pn}{price index for consumer nondurables}
    \item{Ps}{aggregate price index for consumer services}
    \item{Pop}{u.S. total population in millions}
  }
}
\source{
    Consil of Economic Advisors, Economic Report of the President:   1996 ( Washington, D.C.: U.S. Government Printing Office, 1996). \cr
}
\references{
    Greene, W.H. (2003) \emph{Econometric Analysis}, Prentice Hall, Table F2-2. \cr
}
\section{Original description}{``36 yearly observations, 1960-1995.''}
\keyword{datasets}