\name{Income}
\docType{data}
\alias{Income}
\title{Income and Expenditure Data}
\description{
  a cross-section

 \emph{number of observations} :  100

 \emph{observation} :  individuals

   \emph{country} :  United States

}
\usage{data(Income)}
\format{A dataframe containing :
  \describe{
    \item{MDR}{number of derogatory reports}
    \item{Acc}{credit card application accepted (1=yes)}
    \item{Age}{age in years +12ths of a year}
    \item{Income}{income, divided by 10,000}
    \item{Avgexp}{avg. monthly credit card expenditure}
    \item{Ownrent}{individuals owns (1) or rents (0) home}
    \item{Selfempl}{self employed (1=yes,0=no)}
  }
}
\source{
    Greene, W.H. (1992) \emph{A statistical model for credit scoring}, Working Paper No Ec-92-29, New York University, Departement of Economics, Stern School of Business .
}
\references{
    Greene, W.H. (2003) \emph{Econometric Analysis}, Prentice Hall, Table F9.1.
}
\section{Original description}{\dQuote{100 cross-section observations.}}
\keyword{datasets}