\name{Inflation}
\docType{data}
\alias{Inflation}
\title{Inflation Data}
\description{
  quarterly observations from 1972 to 1981\cr

 \emph{number of observations} :  54 \cr
   \emph{observation} :  country \cr
   \emph{country} :  United Kingdown \cr
    }
\usage{data(Inflation)}
\format{A dataframe containing :
  \describe{
    \item{Pa}{actual inflation}
    \item{Pe}{expected inflation}
  }
}
\source{
    Perasan, M.  and  A.  Hall (1988) ``Tests of non-nested linear regression models subject to linear restrictions'', \emph{Economic Letters }, \bold{27}, 341-348. \cr
}
\references{
    Greene, W.H. (2003) \emph{Econometric Analysis}, Prentice Hall, Table F8.1. \cr
}
\keyword{datasets}