\name{Invest}
\docType{data}
\alias{Invest}
\title{Investment}
\description{
  annual observations from 1968 to 1982\cr

 \emph{number of observations} :  15 \cr
   \emph{observation} :  contry \cr
   \emph{country} :  United States \cr
    }
\usage{data(Invest)}
\format{A dataframe containing :
  \describe{
    \item{Year}{date}
    \item{GNP}{nominal GNP}
    \item{Invest}{nominal investment}
    \item{CPI}{consumer price index}
    \item{Interest}{interest rate}
  }
}
\note{``CPI 1967 is 79.06. The interest rate is the average yearly discount rate at the New York Federal Reserve Bank.''}
\source{
    Economic Report of the President (Washington, D.C.: U.S. Government Printing Office, 1983). \cr
}
\references{
    Greene, W.H. (2003) \emph{Econometric Analysis}, Prentice Hall, Table F3.1. \cr
}
\section{Original description}{``15 yearly observations, 1968-1982.''}
\keyword{datasets}