\name{Klein}
\docType{data}
\alias{Klein}
\title{Klein's Model I}
\description{
  yearly observations from 1920 to 1941

 \emph{number of observations} :  22

 \emph{observation} :  country

   \emph{country} :  United States

}
\usage{data(Klein)}
\format{A dataframe containing :
  \describe{
    \item{Year}{date}
    \item{C}{consumption}
    \item{P}{corporate profits}
    \item{Wp}{private wage bill}
    \item{I}{investment}
    \item{K1}{previous year's capital stock}
    \item{X}{gNP}
    \item{Wg}{government wage bill}
    \item{G}{government spending}
    \item{T}{taxes}
  }
}
\source{
    Klein, L. (1950) \emph{Economic fluctuations in the United States, 1921-1941}, New York, John Wiley and Sons.
}
\references{
    Greene, W.H. (2003) \emph{Econometric Analysis}, Prentice Hall, Table F15.1.
}
\section{Original description}{\dQuote{22 yearly observations, 1920-1941.}}
\keyword{datasets}