\name{Longley}
\docType{data}
\alias{Longley}
\title{The Longley Data}
\description{
  annual observations from 1947 to 1962\cr

 \emph{number of observations} :  16 \cr
   \emph{observation} :  country \cr
   \emph{country} :  United States \cr
    }
\usage{data(Longley)}
\format{A dataframe containing :
  \describe{
    \item{Employ}{employment (1,000s)}
    \item{Price}{gNP deflator}
    \item{GNP}{nominal GNP (millions)}
    \item{Armed}{armed forces}
    \item{Year}{date}
  }
}
\source{
    Longley, J. (1967) ``An appraisal of least squares programs from the point of view of the user'', \emph{Journal of the American Statistical Association }, \bold{62}, 819-841. \cr
}
\references{
    Greene, W.H. (2003) \emph{Econometric Analysis}, Prentice Hall, Table F4.2. \cr
}
\section{Original description}{``16 yearly observations, 1947-1962.''}
\keyword{datasets}