\name{Metal}
\docType{data}
\alias{Metal}
\title{Production for SIC 33}
\description{
  a cross-section\cr

 \emph{number of observations} :  27 \cr
   \emph{observation} :  region \cr
   \emph{country} :  United States \cr
    }
\usage{data(Metal)}
\format{A dataframe containing :
  \describe{
    \item{Obs}{observation}
    \item{ValueAdd}{output}
    \item{Labor}{labor input}
    \item{Capital}{capital input}
  }
}
\note{``Data are per etablisment, labor is a measure of labor input, and capital is the gross value of plant and equipment. A scale factor used to normalize the capital figure in the original study has been omitted. Further details on construction of the data are given in Aigner et al. (1977) and in Hildebrand and Liu (1957).''}
\source{
    Aigner , D. ,  K.   Lovell and  P.  Schmidt (1977) ``Formulation and estimation of stochastic frontier production models'', \emph{Journal of Econometrics }, \bold{6}, 21-37. \cr
    Hildebrand, G.  and  T.  Liu (1957) \emph{Manufacturing production functions in the United States}, Ithaca, N.Y.: Cornell University Press. \cr
}
\references{
    Greene, W.H. (2003) \emph{Econometric Analysis}, Prentice Hall, Table F6.1. \cr
}
\section{Original description}{``Primary metals, 27 statewide observations.''}
\keyword{datasets}