\name{ModeChoice}
\docType{data}
\alias{ModeChoice}
\title{Data to Study Travel Mode Choice}
\description{
  a cross-section\cr

 \emph{number of observations} :  840 \cr
   \emph{observation} :  individuals \cr
   \emph{country} :  Australia \cr
    }
\usage{data(ModeChoice)}
\format{A dataframe containing :
  \describe{
    \item{Mode}{choice : air, train, bus or car}
    \item{Ttme}{terminal waiting cost time, 0 for car}
    \item{Invc}{in vehicule cost-cost component}
    \item{Invt}{travel time in vehicule}
    \item{GC}{generalized cost measure}
    \item{Hinc}{household income}
    \item{PSize}{party size in mode chosen}
  }
}
\source{
    Green, W.H.  and  D.  Hensher (1997) \emph{Multinomial logit and discret choice models} \emph{in}  Green, W. H. (1997) \emph{LIMDEP version 7.0 user's manual revised}, Plainview, New York econometric software, Inc. . \cr
}
\references{
    Greene, W.H. (2003) \emph{Econometric Analysis}, Prentice Hall, Table F21.2. \cr
}
\section{Original description}{``840 observations, on 4 modes for 210 individuals.''}
\keyword{datasets}