\name{Money}
\docType{data}
\alias{Money}
\title{Money, Output and Price Deflator Data}
\description{
  quarterly observations from 1950-1 to 1983-4\cr

 \emph{number of observations} :  136 \cr
   \emph{observation} :  country \cr
   \emph{country} :  United States \cr
    }
\usage{data(Money)}
\format{A dataframe containing :
  \describe{
    \item{Qtr}{year.Quarter}
    \item{Y}{nominal GNP}
    \item{M1}{m1 measure of money stock}
    \item{P}{implicit price deflator for GNP}
  }
}
\source{
    National income and product accounts, U.S. Department of Commerce, Bureau of Economic Analysis, Survey of Current Business: Business Statistics. \cr
}
\references{
    Greene, W.H. (2003) \emph{Econometric Analysis}, Prentice Hall, Table F20.2. \cr
}
\section{Original description}{``136 quaterly data observations, 1950-1983.''}
\keyword{datasets}