\name{Mro}
\docType{data}
\alias{Mro}
\title{Labor supply data}
\description{
  a cross-section\cr

 \emph{number of observations} :  753 \cr
   \emph{observation} :  individuals \cr
   \emph{country} :  United States \cr
    }
\usage{data(Mro)}
\format{A dataframe containing :
  \describe{
    \item{LFP}{a dummy variable = 1 if woman worked in 1975, else 0}
    \item{WHRS}{wife's hours of work in 1975}
    \item{KL6}{number of children less than 6 years old in household}
    \item{K618}{number of children between ages 6 and 18 in household}
    \item{WA}{wife's age}
    \item{WE}{wife's educational attainment, in years}
    \item{WW}{wife's average hourly earnings, in 1975 dollars}
    \item{RPWG}{wife's wage reported at the time of the 1976 interview  (not= 1975 estimated wage)}
    \item{HHRS}{husband's hours worked in 1975}
    \item{HA}{husband's age}
    \item{HE}{husband's educational attainment, in years}
    \item{HW}{husband's wage, in 1975 dollars}
    \item{FAMINC}{family income, in 1975 dollars}
    \item{MTR}{nA}
    \item{WMED}{wife's mother's educational attainment, in years}
    \item{WFED}{wife's father's educational attainment, in years}
    \item{UN}{unemployment rate in county of residence, in percentage points}
    \item{CIT}{dummy variable = one if live in large city (SMSA), else zero}
    \item{AX}{actual years of wife's previous labor market experience}
  }
}
\source{
    Mroz, T. (1987) ``The sensivity of an empirical model of married women's hours of work to economic and statistical assumptions'', \emph{Econometrica }, \bold{55}, 765-799. \cr
    1976 Panel Study of Income Dynamics. \cr
}
\references{
    Greene, W.H. (2003) \emph{Econometric Analysis}, Prentice Hall, Table F4.1. \cr
}
\section{Original description}{``from Mroz (1987).''}
\keyword{datasets}