\name{Solow}
\docType{data}
\alias{Solow}
\title{Solow's Technological Change Data}
\description{
  annual observations from 1909 to 1949\cr

 \emph{number of observations} :  41 \cr
   \emph{observation} :  country \cr
   \emph{country} :  United States \cr
    }
\usage{data(Solow)}
\format{A dataframe containing :
  \describe{
    \item{year}{date}
    \item{q}{output}
    \item{k}{capital/labor ratio}
    \item{A}{index of technology}
  }
}
\note{``Several variables are omitted.''}
\source{
    Solow, R. (1957) ``Technical change and the agregate production function'', \emph{Review of Economics and Statistics }, \bold{39}, 312-320. \cr
}
\references{
    Greene, W.H. (2003) \emph{Econometric Analysis}, Prentice Hall, Table F7.2. \cr
}
\section{Original description}{``41 yearly observations, 1909-1949.''}
\keyword{datasets}