\name{War}
\docType{data}
\alias{War}
\title{Consumption and Income}
\description{
  annual observations from 1940 to 1950\cr

 \emph{number of observations} :  11 \cr
   \emph{observation} :  country \cr
   \emph{country} :  United States \cr
    }
\usage{data(War)}
\format{A dataframe containing :
  \describe{
    \item{Year}{date}
    \item{X}{disposable Income}
    \item{C}{consumption}
    \item{W}{war years dummy variable, one in 1942-1945, zero other years}
  }
}
\source{
    Economic Report of the President (Washington D.C.: U.S. Government Printing Office), 1987. \cr
}
\references{
    Greene, W.H. (2003) \emph{Econometric Analysis}, Prentice Hall, Table F2.1. \cr
}
\section{Original description}{``11 yearly observations, 1940-1950.''}
\keyword{datasets}