\name{Bound}
\docType{data}
\alias{Bound}
\title{Bound Yield on a Moody's Aaa Rated}
\description{
  monthly observations from 1990-1 to 1994-12

 \emph{number of observations} :  60

 \emph{observation} :  country

   \emph{country} :  United States

}
\usage{data(Bound)}
\format{A dataframe containing :
  \describe{
    \item{date}{year.Month}
    \item{Y}{bound}
  }
}
\source{
    National income and product accounts, U.S. Department of Commerce, Bureau of Economic Analysis, survey of current business: Business Statistics.
}
\references{
    Greene, W.H. (2003) \emph{Econometric Analysis}, Prentice Hall, Table F20.1.
}
\section{Original description}{\dQuote{60 mounthly observations, 1990-1994.}}
\keyword{datasets}