\name{Conso}
\docType{data}
\alias{Conso}
\title{Consumption and Income}
\description{
  annual observations from 1970 to 1979

 \emph{number of observations} :  10

 \emph{observation} :  Country

   \emph{country} :  United States

}
\usage{data(Conso)}
\format{A dataframe containing :
  \describe{
    \item{C}{consumption}
    \item{Y}{disposable income}
  }
}
\source{
    Concil of Economic Advisors, Economic Report of the President   Washington D.C.: U.S. Government Printing Office, 1987.
}
\references{
    Greene, W.H. (2003) \emph{Econometric Analysis}, Prentice Hall, Table F1.1.
}
\section{Original description}{\dQuote{10 yearly observations, 1970-1979.}}
\keyword{datasets}