\name{EconEduc}
\docType{data}
\alias{EconEduc}
\title{Program Effectiveness}
\description{
  a cross-section

 \emph{number of observations} :  32

 \emph{observation} :  individuals

   \emph{country} :  United States

}
\usage{data(EconEduc)}
\format{A dataframe containing :
  \describe{
    \item{OBS}{observation}
    \item{GPA}{grade point average}
    \item{TUCE}{test score on economic test}
    \item{PSI}{participation in program}
    \item{GRADE}{grade increase (1) or decrease (0) indicator}
  }
}
\source{
    Spector, L.  and  M.  Mazzeo (1980) \dQuote{Probit analysis and economic education}, \emph{Journal of Economic Education }, \bold{11}, 37-44.
}
\references{
    Greene, W.H. (2003) \emph{Econometric Analysis}, Prentice Hall, Table F21.1.
}
\section{Original description}{\dQuote{32 cross section observations.}}
\keyword{datasets}