\name{ExchRate}
\docType{data}
\alias{ExchRate}
\title{Boolerslev and Ghysels Exchange Rate Data}
\description{
  daily observations from  to 

 \emph{number of observations} :  1974

 \emph{observation} :  country

   \emph{country} :  Germany

}
\usage{data(ExchRate)}
\format{A dataframe containing :
  \describe{
    \item{Y}{nominal return on Mark/Pound exchange rate, daily}
  }
}
\source{
    Bollerslev, T. (1986) \dQuote{Generalized autoregressive conditional heteroscedasticity}, \emph{Journal of Econometrics }, \bold{31}, 307-327.
}
\references{
    Greene, W.H. (2003) \emph{Econometric Analysis}, Prentice Hall, Table F11.1.
}
\section{Original description}{\dQuote{1974 daily observations.}}
\keyword{datasets}