\name{Grunfeld}
\docType{data}
\alias{Grunfeld}
\title{Grunfeld Investment Data}
\description{
  a panel of 5 observations  from 1935  to 1954

 \emph{number of observations} :  100

 \emph{observation} :  firms

   \emph{country} :  United States

}
\usage{data(Grunfeld)}
\format{A dataframe containing :
  \describe{
    \item{Year}{date}
    \item{Firm}{observation}
    \item{I}{gross Investment}
    \item{F}{value of the firm}
    \item{C}{stock of plant and equipment}
  }
}
\source{
    Moody's Industrial Manual, Survey of Current Business.
}
\references{
    Greene, W.H. (2003) \emph{Econometric Analysis}, Prentice Hall, Table F13.1.
}
\section{Original description}{\dQuote{100 yearly observations on 5 firms for 1935-1954.}}
\keyword{datasets}