\name{Inflation}
\docType{data}
\alias{Inflation}
\title{Inflation Data}
\description{
  quarterly observations from 1972 to 1981

 \emph{number of observations} :  54

 \emph{observation} :  country

   \emph{country} :  United Kingdown

}
\usage{data(Inflation)}
\format{A dataframe containing :
  \describe{
    \item{Pa}{actual inflation}
    \item{Pe}{expected inflation}
  }
}
\source{
    Perasan, M.  and  A.  Hall (1988) \dQuote{Tests of non-nested linear regression models subject to linear restrictions}, \emph{Economic Letters }, \bold{27}, 341-348.
}
\references{
    Greene, W.H. (2003) \emph{Econometric Analysis}, Prentice Hall, Table F8.1.
}
\keyword{datasets}