\name{Longley}
\docType{data}
\alias{Longley}
\title{The Longley Data}
\description{
  annual observations from 1947 to 1962

 \emph{number of observations} :  16

 \emph{observation} :  country

   \emph{country} :  United States

}
\usage{data(Longley)}
\format{A dataframe containing :
  \describe{
    \item{Employ}{employment (1,000s)}
    \item{Price}{gNP deflator}
    \item{GNP}{nominal GNP (millions)}
    \item{Armed}{armed forces}
    \item{Year}{date}
  }
}
\source{
    Longley, J. (1967) \dQuote{An appraisal of least squares programs from the point of view of the user}, \emph{Journal of the American Statistical Association }, \bold{62}, 819-841.
}
\references{
    Greene, W.H. (2003) \emph{Econometric Analysis}, Prentice Hall, Table F4.2.
}
\section{Original description}{\dQuote{16 yearly observations, 1947-1962.}}
\keyword{datasets}