\name{ManufCost}
\docType{data}
\alias{ManufCost}
\title{Manufacturing Costs}
\description{
  annual observations from 1947 to 1971

 \emph{number of observations} :  25

 \emph{observation} :  country

   \emph{country} :  United States

}
\usage{data(ManufCost)}
\format{A dataframe containing :
  \describe{
    \item{Year}{date}
    \item{Cost}{cost index}
    \item{K}{capital cost share}
    \item{L}{labor cost share}
    \item{E}{energy cost share}
    \item{M}{materials cost share}
    \item{Pk}{capital price}
    \item{Pl}{labor price}
    \item{Pe}{energy price}
    \item{Pm}{materials price}
  }
}
\source{
    Berndt, E.  and  D.  Wood (1975) \dQuote{Technology, prices and the derived demand for energy}, \emph{Journal of Economics and Statistics }, \bold{57}, 376-384.
}
\references{
    Greene, W.H. (2003) \emph{Econometric Analysis}, Prentice Hall, Table F14.1.
}
\section{Original description}{\dQuote{U.S. economy, 25 yearly observations, 1947-1971.}}
\keyword{datasets}