\name{Nerlove}
\docType{data}
\alias{Nerlove}
\title{Cost function, 145 U.S. Electricity Producers}
\description{
  a cross-section from 1955

 \emph{number of observations} :  159

 \emph{observation} :  firms

   \emph{country} :  United States

}
\usage{data(Nerlove)}
\format{A dataframe containing :
  \describe{
    \item{Firm}{observation}
    \item{Year}{1955 for all observations}
    \item{Cost}{total cost}
    \item{Output}{total output}
    \item{Pl}{wage rate}
    \item{Sl}{cost share for labor}
    \item{Pk}{capital price index}
    \item{Sk}{cost share for capital}
    \item{Pf}{fuel price}
    \item{Sf}{cost share for fuel}
  }
}
\note{\dQuote{The data file contains several extra observations that are aggregates of commonly owned firms. Use only the first 145 for analysis.}}
\source{
    Nerlove, M. (1963) \emph{Returns to scale in electricity industry} \emph{in}  Christ, C. ed. (1963) \emph{Measurment in economics: studies in mathematical economics and econometrics in memory of Yehuda Grunfeld }, Standford, California, Standford University Press .
    Christensen, L.  and  W. H.  Greene (1976) \dQuote{Economies of scale in U.S. electric power generation}, \emph{Journal of Political Economy }, \bold{84}, 655-676.
}
\references{
    Greene, W.H. (2003) \emph{Econometric Analysis}, Prentice Hall, Table F14.2.
}
\section{Original description}{\dQuote{1955, data Nerlove.}}
\keyword{datasets}