\name{Solow}
\docType{data}
\alias{Solow}
\title{Solow's Technological Change Data}
\description{
  annual observations from 1909 to 1949

 \emph{number of observations} :  41

 \emph{observation} :  country

   \emph{country} :  United States

}
\usage{data(Solow)}
\format{A dataframe containing :
  \describe{
    \item{year}{date}
    \item{q}{output}
    \item{k}{capital/labor ratio}
    \item{A}{index of technology}
  }
}
\note{\dQuote{Several variables are omitted.}}
\source{
    Solow, R. (1957) \dQuote{Technical change and the agregate production function}, \emph{Review of Economics and Statistics }, \bold{39}, 312-320.
}
\references{
    Greene, W.H. (2003) \emph{Econometric Analysis}, Prentice Hall, Table F7.2.
}
\section{Original description}{\dQuote{41 yearly observations, 1909-1949.}}
\keyword{datasets}