\name{Strike}
\docType{data}
\alias{Strike}
\title{Strike Duration Data}
\description{
  a cross-section from 1968\cr

 \emph{number of observations} :  62 \cr
   \emph{observation} :   \cr
   \emph{country} :  United States \cr
    }
\usage{data(Strike)}
\format{A dataframe containing :
  \describe{
    \item{T}{strike duration in days}
    \item{Prod}{unanticipated output}
  }
}
\source{
    Kennan, J. (1985) ``The duration of contract strikes in U.S. manufacturing'', \emph{Journal of Econometrics }, \bold{28}, 5-28. \cr
}
\references{
    Greene, W.H. (2003) \emph{Econometric Analysis}, Prentice Hall, Table F22.1. \cr
}
\section{Original description}{``62 observations in 9 years, 1968-1976.''}
\keyword{datasets}