\name{TranspEq}
\docType{data}
\alias{TranspEq}
\title{Statewide Data on Transportation Equipment Manufacturing}
\description{
  a cross-section

 \emph{number of observations} :  25

 \emph{observation} :  region

   \emph{country} :  United States

}
\usage{data(TranspEq)}
\format{A dataframe containing :
  \describe{
    \item{State}{observation}
    \item{ValueAdd}{output}
    \item{Capital}{capital input}
    \item{Labor}{labor input}
    \item{NFirm}{number of firms}
  }
}
\note{\dQuote{\dQuote{Value added}, \dQuote{Capital} and \dQuote{Labor} are in millions of 1957 dollars. Data used for regression examples are per etablishment. Raw data are used in stochastic frontier application in chapter 16.}}
\source{
    Zellner, A.  and  N.  Revankar (1970) \dQuote{Generalized production functions}, \emph{Review of Economic Studies }, \bold{37}, 241-250.
}
\references{
    Greene, W.H. (2003) \emph{Econometric Analysis}, Prentice Hall, Table F9.2.
}
\section{Original description}{\dQuote{25 Observations.}}
\keyword{datasets}